Add hooks to enable compiling scss into jar/uberjar
Leiningen has hook functionality that enable plugins to tie into the
existing tasks of lein. lein-scss seems to be the only scss plugin that does
not leverage this, which is unfortunate since it seems to be the only one
that can use the native libsass library for compilation. This commit should
provide the neccessary hook functions that will expose both compile and jar
tasks to leiningen so that a single call to `lein compile` will output scss
as the project is configured, and calling `lein jar` or `lein uberjar` will
enable adding compiled scss as css to the resulting jarfile.

In order to better suit users, I've added a :jar keyword flag to the build
profile of the scss configuration. Setting this to false will prevent the
process from adding the css into the jarfile.

https://github.com/technomancy/leiningen/blob/master/doc/PLUGINS.md#hooks